% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment1.R
\docType{package}
\name{alignment1}
\alias{alignment1}
\alias{alignment1-package}
\title{Alignment 1 experiment}
\description{
Data and materials for the Alignment 1 experiment.
}
\details{
alignment1
A data package for the Alignment 1 experiment.

Use \code{data(package='alignment1')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
