% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment1.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{id}{}
\item{alignment}{}
\item{reliability_amount}{}
\item{sex}{}
\item{age}{}
\item{project}{}
\item{allocation}{}
\item{confidence}{}
\item{npv_amount}{}
\item{sample}{}
\item{max_npv}{}
\item{min_npv}{}
\item{max_min_difference}{}
}
}
\source{
The data was collected using Qualtrics.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment1}
}
