% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment1.R
\docType{package}
\name{alignment1}
\alias{alignment1}
\alias{alignment1-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
alignment1
A data package for alignment1.

Use \code{data(package='alignment1')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
